todo-app/requirements.txt
asgiref==3.3.1
Django==3.1.7
djangorestframework==3.12.2
djangorestframework-simplejwt==4.6.0
psycopg2==2.8.6
PyJWT==2.0.1
python-dotenv==0.15.0
pytz==2021.1
sqlparse==0.4.1
